vuedjango/requirements.txt
django == 2.1
djangorestframework == 3.9.0
django_cors_headers == 2.4.0